完善专栏功能

+++ src/main/java/com/wblog/model/vo/ColumnItemVo.java
@@ -36,5 +36,5 @@ public class ColumnItemVo implements Serializable {
     /**
      * 收藏人数
      */
-    private Integer collectNum;
+    private Long collectNum;
 }
